impr(sdk js standalone): add notes about authentication
This commit adds a note about the fact that you have to be authenticated
in order to use the example correctly.

It also display the task name in the list (rather than its ID) when
available and fixes some typo and changes the service endpoint to
avoid authentication issues.

Related to: CAM-3114